requirements.txt
requests
python-owasp-zap-v2.4
python-nmap
tqdm
Jinja2
